#22 Rename module (dataspec → specs)
