+++ pandas_dataclasses/__init__.py
@@ -4,12 +4,12 @@ __version__ = "0.1.0"
 
 
 # submodules
-from . import dataspec
 from . import series
+from . import specs
 from . import typing
 
 
 # aliases
-from .dataspec import *
 from .series import *
+from .specs import *
 from .typing import *
